Tool Mode일경우, F키 더블탭 - Tool Camera를 오브젝트쪽으로 이동, F + 마우스 왼쪽 - 마우스 위치로 오브젝트 이동, M + 마우스 왼쪽 - offset 값으로 오브젝트 이동

+++ Project/Client/Animator2DEditorGUI.cpp
@@ -508,7 +508,7 @@ void Animator2DEditorGUI::_FixedAnimationPanel()
 	
 
 	// frame Á¶Àý
-	ImGui::SliderInt("Frame## FixedAnimator2D", &m_iCurFrameIdx, 0, max(0, vecAnimFrm.size() - 1));
+	ImGui::SliderInt("Frame## FixedAnimator2D", &m_iCurFrameIdx, 0, max(0, (int)(vecAnimFrm.size() - 1)));
 
 	CTexture* pCurTex = pCurAnim->GetTextures()[m_iCurFrameIdx].Get();
